+++ update.sh
@@ -1,0 +1,17 @@
+yarn upgrade \
+@tiptap/core@next \
+@tiptap/extension-code@next \
+@tiptap/extension-code-block@next \
+@tiptap/extension-color@next \
+@tiptap/extension-highlight@next \
+@tiptap/extension-image@next \
+@tiptap/extension-link@next \
+@tiptap/extension-subscript@next \
+@tiptap/extension-superscript@next \
+@tiptap/extension-table@next \
+@tiptap/extension-text-align@next \
+@tiptap/extension-text-style@next \
+@tiptap/extension-underline@next \
+@tiptap/extensions@next \
+@tiptap/pm@next \
+@tiptap/starter-kit@next